bug: nm-wait failes on rebuild
- This is related to udev not initializing
devices. NetworkManager never completes startup
because a WireGuard (tailscale) interface is
never initialized by udev. A workaround is just
putting the affected device into the
unmanaged list.
- I had seen other code samples where docker0
was also added to the unmanaged list.

On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
	modified:   modules/hw/networking/default.nix